Change application name to 'QuiCSV'
'CSVedit' was literally the first thing I thought of when I just needed
a name for the folder to store the code. An itch at the back of my
brain told me to search github before I uploaded anything, and good
thing, too--otherwise I may not have known about the 200 (or so) other
projects called "CSVedit" (or something very similar), and would have
falsely believed myself a bastion of original and iconoclastic thinking.

Anyway, CSVedit was lame. QuiCSV may not be much better, I don't know.
It's at least a little different. And starts with Q! Since it, uh, uses
Qt. So there's that. And yes, it's pronounced "Quick See Ess Vee",
because I said so.